+++ src/reference/corporate.rs
@@ -11,10 +11,10 @@ use crate::{
     deserialize::{deserialize_date_time, deserialize_opt_date_time_hash_map},
     historical::{handle_zstd_jsonl_response, AddToForm, ReqwestForm},
     reference::{
-        Action, Country, Currency, DateTimeLike, End, Event, EventSubType, Fraction, GlobalStatus,
-        ListingSource, ListingStatus, MandVolu, OutturnStyle, PaymentType, SecurityType, Start,
+        Action, Country, Currency, End, Event, EventSubType, Fraction, GlobalStatus, ListingSource,
+        ListingStatus, MandVolu, OutturnStyle, PaymentType, SecurityType, Start,
     },
-    Symbols,
+    DateTimeLike, Symbols,
 };
 
 /// A client for the corporate actions group of Reference API endpoints.